cdk/requirements.txt
aws-cdk.core==1.95.0
aws-cdk.aws_autoscaling==1.95.0
aws-cdk.aws_ec2==1.95.0
aws-cdk.aws_ecs==1.95.0
aws-cdk.aws_lambda==1.95.0
aws-cdk.aws_s3==1.95.0
aws-cdk.aws_lambda_event_sources==1.95.0
aws-cdk.aws_iam==1.95.0